Created 2 separate jobs to run the sm and pm tests
