+++ test/e2e/mount_test.go
@@ -1,4 +1,4 @@
-// Copyright 2020 Google LLC
+// Copyright 2025 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,6 +11,10 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+//go:build secretmanager_e2e || parametermanager_e2e || all_e2e
+// +build secretmanager_e2e parametermanager_e2e all_e2e
+
 package test
 
 import (
@@ -43,7 +47,12 @@ type testFixture struct {
 	secretStoreVersion  string
 	gkeVersion          string
 	location            string
+
 	// below fields explicitly used for parameter manager
+	pmReferenceGlobalSecret1       string
+	pmReferenceGlobalSecret2       string
+	pmReferenceRegionalSecret1     string
+	pmReferenceRegionalSecret2     string
 	parameterIdYaml                string
 	parameterIdJson                string
 	parameterVersionIdYAML         string
@@ -87,44 +96,6 @@ func checkMountedSecret(secretId string) error {
 	}
 	if !bytes.Equal(stdout.Bytes(), []byte(secretId)) {
 		return fmt.Errorf("Secret value is %v, want: %v", stdout.String(), secretId)
-	}
-	return nil
-}
-
-func checkMountedParameterVersion(podName, filePath, expectedPayload string) error {
-	var stdout, stderr bytes.Buffer
-	command := exec.Command("kubectl", "exec", podName,
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default",
-		"--",
-		"cat", filePath)
-	command.Stdout = &stdout
-	command.Stderr = &stderr
-	if err := command.Run(); err != nil {
-		fmt.Println("Stdout:", stdout.String())
-		fmt.Println("Stderr:", stderr.String())
-		return fmt.Errorf("could not read parameter version from container: %w", err)
-	}
-	if !bytes.Equal(stdout.Bytes(), []byte(expectedPayload)) {
-		return fmt.Errorf("parameter version payload value is %v, want: %v", stdout.String(), expectedPayload)
-	}
-	return nil
-}
-
-func checkMountedParameterVersionFileMode(dataFilePath, fileMode string) error {
-	var stdout, stderr bytes.Buffer
-	command := exec.Command("kubectl", "exec", "test-parameter-version-mounter-filemode",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default",
-		"--",
-		"stat", "--printf", "%a", dataFilePath)
-	command.Stdout = &stdout
-	command.Stderr = &stderr
-	if err := command.Run(); err != nil {
-		fmt.Println("Stdout:", stdout.String())
-		fmt.Println("Stderr:", stderr.String())
-		return fmt.Errorf("could not read parameter version file %s from container, error: %w", dataFilePath, err)
-	}
-	if !bytes.Equal(stdout.Bytes(), []byte(fileMode)) {
-		return fmt.Errorf("parameter version file mode is %v, want: %s", stdout.String(), fileMode)
 	}
 	return nil
 }
@@ -180,29 +151,6 @@ func replaceTemplate(templateFile string, destFile string) error {
 	return os.WriteFile(destFile, []byte(template), 0644)
 }
 
-// getParameterPrincipalID describes a parameter and returns its iamPolicyUidPrincipal.
-func getParameterPrincipalID(parameterID, location, projectID string) (string, error) {
-	var stdout, stderr bytes.Buffer
-	args := []string{
-		"parametermanager", "parameters", "describe", parameterID,
-		"--location", location,
-		"--project", projectID,
-		"--format=value(policyMember.iamPolicyUidPrincipal)",
-	}
-	log.Println("+ gcloud", strings.Join(args, " ")) // Log the command
-	cmd := exec.Command("gcloud", args...)
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-
-	err := cmd.Run()
-	if err != nil {
-		log.Fatalf("Stdout: %v\n", stdout.String())
-		log.Fatalf("Stderr: %v\n", stderr.String())
-		return "", fmt.Errorf("failed to describe parameter %s in location %s: %w\nStderr: %s", parameterID, location, err, stderr.String())
-	}
-	return strings.TrimSpace(stdout.String()), nil
-}
-
 // countGcloudVersions lists versions for a secret and returns the count or an error.
 func countGcloudVersions(secretID, projectID, locationID string) (int, error) {
 	args := []string{"secrets", "versions", "list", secretID, "--project", projectID, "--format=value(name)"}
@@ -253,7 +201,7 @@ func waitForMinVersions(t *testing.T, secretID, projectID, locationID string, mi
 }
 
 // Executed before any tests are run. Setup is only run once for all tests in the suite.
-func setupTestSuite(isTokenPassed bool) {
+func setupTestSuite(isTokenPassed bool, suiteType string) {
 
 	rand.Seed(time.Now().UTC().UnixNano())
 
@@ -266,7 +214,7 @@ func setupTestSuite(isTokenPassed bool) {
 		log.Fatal("PROJECT_ID is empty")
 	}
 	f.location = os.Getenv("LOCATION_ID")
-	if len(f.testProjectID) == 0 {
+	if len(f.location) == 0 {
 		log.Fatal("LOCATION_ID is empty")
 	}
 	f.secretStoreVersion = os.Getenv("SECRET_STORE_VERSION")
@@ -292,20 +240,6 @@ func setupTestSuite(isTokenPassed bool) {
 	check(err)
 	f.tempDir = tempDir
 	f.testClusterName = fmt.Sprintf("testcluster-%d", rand.Int31())
-	f.testSecretID = fmt.Sprintf("testsecret-%d", rand.Int31())
-
-	f.testRotateSecretID = f.testSecretID + "-rotate"
-	f.testExtractSecretID = f.testSecretID + "-extract"
-
-	// Parameter manager specific e2e fields
-	f.parameterIdYaml = fmt.Sprintf("testparameteryaml-%d", rand.Int31())
-	f.parameterIdJson = fmt.Sprintf("testparameterjson-%d", rand.Int31())
-	f.parameterVersionIdYAML = fmt.Sprintf("testparameterversionyaml-%d", rand.Int31())
-	f.parameterVersionIdJSON = fmt.Sprintf("testparameterversionjson-%d", rand.Int31())
-	f.regionalParameterIdYAML = fmt.Sprintf("testregionalparameteryaml-%d", rand.Int31())
-	f.regionalParameterIdJSON = fmt.Sprintf("testregionalparameterjson-%d", rand.Int31())
-	f.regionalParameterVersionIdYAML = fmt.Sprintf("testregionalparameterversion-%d", rand.Int31())
-	f.regionalParameterVersionIdJSON = fmt.Sprintf("testregionalparameterversionjson-%d", rand.Int31())
 
 	// Build the plugin deploy yaml
 	pluginFile := filepath.Join(tempDir, "provider-gcp-plugin.yaml")
@@ -339,114 +273,14 @@ func setupTestSuite(isTokenPassed bool) {
 	check(execCmd(exec.Command("kubectl", "apply", "--kubeconfig", f.kubeconfigFile,
 		"-f", pluginFile)))
 
-	// Create test secret
-	secretFile := filepath.Join(f.tempDir, "secretValue")
-	check(os.WriteFile(secretFile, []byte(f.testSecretID), 0644))
-	check(execCmd(exec.Command("gcloud", "secrets", "create", f.testSecretID, "--replication-policy", "automatic",
-		"--data-file", secretFile, "--project", f.testProjectID)))
+	// Create f.testSecretID (for direct SM tests) if secretmanager or all suite is run
+	if suiteType == "secretmanager" || suiteType == "all" {
+		setupSmTestSuite()
+	}
 
-	// Create test parameter and parameter versions -> global region (both YAML and JSON)
-	parameterVersionFileYaml := filepath.Join(f.tempDir, "parameterValueYaml.yaml")
-	parameterVersionFileJson := filepath.Join(f.tempDir, "parameterValueJson.json")
-
-	// Write the byte payload of the parameters into files similar to how secret manager is doing it.
-	check(os.WriteFile(parameterVersionFileYaml, []byte(
-		fmt.Sprintf("user: admin\nuser2: support\ndb_pwd: __REF__(//secretmanager.googleapis.com/projects/%s/secrets/%s/versions/1)\n",
-			f.testProjectID, f.testSecretID)), 0644))
-	check(os.WriteFile(parameterVersionFileJson, []byte(
-		fmt.Sprintf("{\"user\": \"admin\",\n \"user2\": \"support\", \"db_pwd\": \"__REF__(//secretmanager.googleapis.com/projects/%s/secrets/%s/versions/1)\"}",
-			f.testProjectID, f.testSecretID)), 0644))
-
-	// Create Parameters first
-	check(execCmd(exec.Command("gcloud", "parametermanager", "parameters", "create", f.parameterIdYaml,
-		"--location", "global", "--parameter-format", "YAML", "--project", f.testProjectID)))
-
-	check(execCmd(exec.Command("gcloud", "parametermanager", "parameters", "create", f.parameterIdJson,
-		"--location", "global", "--parameter-format", "JSON", "--project", f.testProjectID)))
-
-	// Grant parameter principals access to the global secret
-	globalYamlPrincipal, err := getParameterPrincipalID(f.parameterIdYaml, "global", f.testProjectID)
-	check(err) // Use check(err) which panics on error
-	check(execCmd(exec.Command("gcloud", "secrets", "add-iam-policy-binding", f.testSecretID,
-		"--member", globalYamlPrincipal,
-		"--role", "roles/secretmanager.secretAccessor",
-		"--project", f.testProjectID)))
-
-	globalJsonPrincipal, err := getParameterPrincipalID(f.parameterIdJson, "global", f.testProjectID)
-	check(err)
-	check(execCmd(exec.Command("gcloud", "secrets", "add-iam-policy-binding", f.testSecretID,
-		"--member", globalJsonPrincipal,
-		"--role", "roles/secretmanager.secretAccessor",
-		"--project", f.testProjectID)))
-
-	// Now create the versions using the files you just wrote
-	check(execCmd(exec.Command("gcloud", "parametermanager", "parameters", "versions", "create", f.parameterVersionIdYAML,
-		"--parameter", f.parameterIdYaml, "--location", "global",
-		"--payload-data-from-file", parameterVersionFileYaml, // Use the file path here
-		"--project", f.testProjectID)))
-
-	check(execCmd(exec.Command("gcloud", "parametermanager", "parameters", "versions", "create", f.parameterVersionIdJSON,
-		"--parameter", f.parameterIdJson, "--location", "global",
-		"--payload-data-from-file", parameterVersionFileJson, // And here
-		"--project", f.testProjectID)))
-
-	// Create regional secret
-	secretFile = filepath.Join(f.tempDir, "regionalSecretValue")
-	check(os.WriteFile(secretFile, []byte(f.testSecretID+"-regional"), 0644))
-	check(execCmd(exec.Command("gcloud", "config", "set", "api_endpoint_overrides/secretmanager",
-		"https://secretmanager."+f.location+".rep.googleapis.com/")))
-	check(execCmd(exec.Command("gcloud", "secrets", "create", f.testSecretID, "--location", f.location,
-		"--data-file", secretFile, "--project", f.testProjectID)))
-
-	// Create regional parameter and regional parameter version
-	parameterVersionFileYamlRegional := filepath.Join(f.tempDir, "parameterValueYamlRegional.yaml")
-	parameterVersionFileJsonRegional := filepath.Join(f.tempDir, "parameterValueJsonRegional.json")
-	check(os.WriteFile(parameterVersionFileYamlRegional, []byte(
-		fmt.Sprintf("user: admin\nuser2: support\ndb_pwd: __REF__(//secretmanager.googleapis.com/projects/%s/locations/%s/secrets/%s/versions/1)\n",
-			f.testProjectID, f.location, f.testSecretID)), 0644))
-	check(os.WriteFile(parameterVersionFileJsonRegional, []byte(
-		fmt.Sprintf("{\"user\": \"admin\",\n \"user2\": \"support\", \"db_pwd\": \"__REF__(//secretmanager.googleapis.com/projects/%s/locations/%s/secrets/%s/versions/1)\"}",
-			f.testProjectID, f.location, f.testSecretID)), 0644))
-
-	// Set regional endpoint
-	check(execCmd(exec.Command("gcloud", "config", "set", "api_endpoint_overrides/parametermanager",
-		"https://parametermanager."+f.location+".rep.googleapis.com/")))
-
-	// Create regional YAML and JSON parameters.
-	check(execCmd(exec.Command("gcloud", "parametermanager", "parameters", "create", f.regionalParameterIdYAML,
-		"--location", f.location, "--parameter-format", "YAML", "--project", f.testProjectID)))
-	check(execCmd(exec.Command("gcloud", "parametermanager", "parameters", "create", f.regionalParameterIdJSON,
-		"--location", f.location, "--parameter-format", "JSON", "--project", f.testProjectID)))
-
-	// Grant parameter principals access to the regional secret
-	regionalYamlPrincipal, err := getParameterPrincipalID(f.regionalParameterIdYAML, f.location, f.testProjectID)
-	check(err)
-	check(execCmd(exec.Command("gcloud", "secrets", "add-iam-policy-binding", f.testSecretID, // Assuming regional secret has same ID but different location context
-		"--member", regionalYamlPrincipal,
-		"--role", "roles/secretmanager.secretAccessor",
-		"--project", f.testProjectID, "--location", f.location))) // Need location for regional secret binding
-
-	regionalJsonPrincipal, err := getParameterPrincipalID(f.regionalParameterIdJSON, f.location, f.testProjectID)
-	check(err)
-	check(execCmd(exec.Command("gcloud", "secrets", "add-iam-policy-binding", f.testSecretID, // Assuming regional secret has same ID but different location context
-		"--member", regionalJsonPrincipal,
-		"--role", "roles/secretmanager.secretAccessor",
-		"--project", f.testProjectID, "--location", f.location))) // Need location for regional secret binding
-
-	// Now create corresponding parameter versions to YAML and JSON parameters just created
-	check(execCmd(exec.Command("gcloud", "parametermanager", "parameters", "versions", "create", f.regionalParameterVersionIdYAML,
-		"--parameter", f.regionalParameterIdYAML, "--location", f.location,
-		"--payload-data-from-file", parameterVersionFileYamlRegional, // Use the file path here
-		"--project", f.testProjectID)))
-
-	check(execCmd(exec.Command("gcloud", "parametermanager", "parameters", "versions", "create", f.regionalParameterVersionIdJSON,
-		"--parameter", f.regionalParameterIdJSON, "--location", f.location,
-		"--payload-data-from-file", parameterVersionFileJsonRegional, // And here
-		"--project", f.testProjectID)))
-
-	// Setting endpoints back to the global defaults
-	check(execCmd(exec.Command("gcloud", "config", "unset", "api_endpoint_overrides/secretmanager")))
-	check(execCmd(exec.Command("gcloud", "config", "unset", "api_endpoint_overrides/parametermanager")))
+	if suiteType == "parametermanager" || suiteType == "all" {
+		setupPmTestSuite()
+	}
 
 	if isTokenPassed {
 		type metadataStruct struct {
@@ -511,7 +345,7 @@ func setupTestSuite(isTokenPassed bool) {
 }
 
 // Executed after tests are run. Teardown is only run once for all tests in the suite.
-func teardownTestSuite() {
+func teardownTestSuite(suiteType string) {
 	// print cluster information, useful when debugging
 	execCmd(exec.Command(
 		"kubectl", "describe", "pods",
@@ -534,101 +368,29 @@ func teardownTestSuite() {
 	// Cleanup
 	os.RemoveAll(f.tempDir)
 	execCmd(exec.Command("kubectl", "delete", "containercluster", f.testClusterName))
-	execCmd(exec.Command(
-		"gcloud", "secrets", "delete", f.testSecretID,
-		"--project", f.testProjectID,
-		"--quiet",
-	))
-	execCmd(exec.Command(
-		"gcloud", "secrets", "delete", f.testRotateSecretID,
-		"--project", f.testProjectID,
-		"--quiet",
-	))
-	execCmd(exec.Command(
-		"gcloud", "secrets", "delete", f.testExtractSecretID,
-		"--project", f.testProjectID,
-		"--quiet",
-	))
 
-	// Execute gcloud delete parameter version and delete parameter -> Both YAML and JSON
-	// gcloud parametermanager parameters versions delete PARAMETER_VERSION_ID --parameter=PARAMETER_ID --location=global
-	check(execCmd(exec.Command(
-		"gcloud", "parametermanager", "parameters", "versions", "delete", f.parameterVersionIdYAML,
-		"--parameter", f.parameterIdYaml,
-		"--location", "global",
-		"--project", f.testProjectID,
-		"--quiet",
-	)))
-	check(execCmd(exec.Command(
-		"gcloud", "parametermanager", "parameters", "versions", "delete", f.parameterVersionIdJSON,
-		"--parameter", f.parameterIdJson,
-		"--location", "global",
-		"--project", f.testProjectID,
-		"--quiet",
-	)))
-	// gcloud parametermanager parameters delete PARAMETER_ID --location=global
-	check(execCmd(exec.Command(
-		"gcloud", "parametermanager", "parameters", "delete", f.parameterIdYaml,
-		"--location", "global",
-		"--project", f.testProjectID,
-		"--quiet",
-	)))
-	check(execCmd(exec.Command(
-		"gcloud", "parametermanager", "parameters", "delete", f.parameterIdJson,
-		"--location", "global",
-		"--project", f.testProjectID,
-		"--quiet",
-	)))
-	// gcloud parametermanager parameters delete PARAMETER_ID --location=global
+	if suiteType == "secretmanager" || suiteType == "all" {
+		teardownSmTestSuite()
+	}
 
-	// Cleanup regional secret
-	check(execCmd(exec.Command("gcloud", "config", "set", "api_endpoint_overrides/secretmanager",
-		"https://secretmanager."+f.location+".rep.googleapis.com/")))
-	check(execCmd(exec.Command("gcloud", "secrets", "delete", f.testSecretID, "--location", f.location,
-		"--project", f.testProjectID, "--quiet")))
-	check(execCmd(exec.Command("gcloud", "secrets", "delete", f.testRotateSecretID, "--location", f.location,
-		"--project", f.testProjectID, "--quiet")))
-	check(execCmd(exec.Command("gcloud", "config", "unset", "api_endpoint_overrides/secretmanager")))
-
-	// Clean regional parameters -> Both YAML and JSON
-	check(execCmd(exec.Command("gcloud", "config", "set", "api_endpoint_overrides/parametermanager",
-		"https://parametermanager."+f.location+".rep.googleapis.com/")))
-
-	check(execCmd(exec.Command(
-		"gcloud", "parametermanager", "parameters", "versions", "delete", f.regionalParameterVersionIdYAML,
-		"--parameter", f.regionalParameterIdYAML,
-		"--location", f.location,
-		"--project", f.testProjectID,
-		"--quiet",
-	)))
-	check(execCmd(exec.Command(
-		"gcloud", "parametermanager", "parameters", "versions", "delete", f.regionalParameterVersionIdJSON,
-		"--parameter", f.regionalParameterIdJSON,
-		"--location", f.location,
-		"--project", f.testProjectID,
-		"--quiet",
-	)))
-
-	check(execCmd(exec.Command(
-		"gcloud", "parametermanager", "parameters", "delete", f.regionalParameterIdYAML,
-		"--location", f.location,
-		"--project", f.testProjectID,
-		"--quiet",
-	)))
-	check(execCmd(exec.Command(
-		"gcloud", "parametermanager", "parameters", "delete", f.regionalParameterIdJSON,
-		"--location", f.location,
-		"--project", f.testProjectID,
-		"--quiet",
-	)))
-
-	//gcloud parametermanager parameters versions delete PARAMETER_VERSION_ID --parameter=PARAMETER_IDl --location=LOCATION
-	// gcloud parametermanager parameters delete PARAMETER_ID --location=LOCATION
-	check(execCmd(exec.Command("gcloud", "config", "unset", "api_endpoint_overrides/parametermanager")))
+	if suiteType == "parametermanager" || suiteType == "all" {
+		teardownPmTestSuite()
+	}
 }
 
 // Entry point for go test.
 func TestMain(m *testing.M) {
+
+	suiteType := os.Getenv("E2E_TEST_SUITE")
+	if suiteType == "" {
+		// This default is for local runs or if the e2e_all.test binary is run
+		// without E2E_TEST_SUITE explicitly set to "all".
+		// The setup/teardown logic for "all" will handle both SM and PM resources.
+		log.Println("E2E_TEST_SUITE environment variable not set, defaulting to 'all' for setup/teardown logic.")
+		suiteType = "all"
+	}
+	log.Printf("Running E2E tests for suite: %s\n", suiteType)
+
 	withoutTokenStatus := runTest(m, false)
 	withTokenStatus := runTest(m, true)
 	fmt.Printf("Exit Code when token is not passed from driver to provder is: %v\n", withoutTokenStatus)
@@ -637,517 +399,15 @@ func TestMain(m *testing.M) {
 }
 
 // Handles setup/teardown test suite and runs test. Returns exit code.
-func runTest(m *testing.M, isTokenPassed bool) (code int) {
+func runTest(m *testing.M, isTokenPassed bool, suiteType string) (code int) {
 	defer func() {
 		if r := recover(); r != nil {
 			fmt.Println("Test execution panic:", r)
 			code = 1
 		}
-		teardownTestSuite()
+		teardownTestSuite(suiteType)
 	}()
 
-	setupTestSuite(isTokenPassed)
+	setupTestSuite(isTokenPassed, suiteType)
 	return m.Run()
-}
-
-// Execute a test job that mounts a secret and checks that the value is correct.
-func TestMountSecret(t *testing.T) {
-	podFile := filepath.Join(f.tempDir, "test-pod.yaml")
-	if err := replaceTemplate("templates/test-pod.yaml.tmpl", podFile); err != nil {
-		t.Fatalf("Error replacing pod template: %v", err)
-	}
-
-	if err := execCmd(exec.Command("kubectl", "apply", "--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default", "-f", podFile)); err != nil {
-		t.Fatalf("Error creating job: %v", err)
-	}
-
-	// As a workaround for https://github.com/kubernetes/kubernetes/issues/83242, we sleep to
-	// ensure that the job resources exists before attempting to wait for it.
-	time.Sleep(5 * time.Second)
-	if err := execCmd(exec.Command("kubectl", "wait", "pod/test-secret-mounter", "--for=condition=Ready",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default", "--timeout", "5m")); err != nil {
-		t.Fatalf("Error waiting for job: %v", err)
-	}
-
-	if err := checkMountedSecret(f.testSecretID); err != nil {
-		t.Fatalf("Error while testing global secret: %v", err)
-	}
-	if err := checkMountedSecret(f.testSecretID + "-regional"); err != nil {
-		t.Fatalf("Error while testing regional secret: %v", err)
-	}
-}
-
-func TestMountSecretFileMode(t *testing.T) {
-	podFile := filepath.Join(f.tempDir, "test-mode.yaml")
-	if err := replaceTemplate("templates/test-mode.yaml.tmpl", podFile); err != nil {
-		t.Fatalf("Error replacing mode template: %v", err)
-	}
-
-	if err := execCmd(exec.Command("kubectl", "apply", "--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default", "-f", podFile)); err != nil {
-		t.Fatalf("Error creating job: %v", err)
-	}
-
-	// As a workaround for https://github.com/kubernetes/kubernetes/issues/83242, we sleep to
-	// ensure that the job resources exists before attempting to wait for it.
-	time.Sleep(5 * time.Second)
-	if err := execCmd(exec.Command("kubectl", "wait", "pod/test-secret-mode", "--for=condition=Ready",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default", "--timeout", "5m")); err != nil {
-		t.Fatalf("Error waiting for job: %v", err)
-	}
-
-	// stat the file in the symlinked '..data' directory, symlink will always return 777 otherwise
-	if err := checkFileMode(f.testSecretID); err != nil {
-		t.Fatalf("Error while testing global secret: %v", err)
-	}
-	if err := checkFileMode(f.testSecretID + "-regional"); err != nil {
-		t.Fatalf("Error while testing regional secret: %v", err)
-	}
-}
-
-func TestMountNestedPath(t *testing.T) {
-	podFile := filepath.Join(f.tempDir, "test-nested.yaml")
-	if err := replaceTemplate("templates/test-nested.yaml.tmpl", podFile); err != nil {
-		t.Fatalf("Error replacing pod template: %v", err)
-	}
-
-	if err := execCmd(exec.Command("kubectl", "apply", "--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default", "-f", podFile)); err != nil {
-		t.Fatalf("Error creating job: %v", err)
-	}
-
-	// As a workaround for https://github.com/kubernetes/kubernetes/issues/83242, we sleep to
-	// ensure that the job resources exists before attempting to wait for it.
-	time.Sleep(5 * time.Second)
-	if err := execCmd(exec.Command("kubectl", "wait", "pod/test-secret-nested", "--for=condition=Ready",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default", "--timeout", "5m")); err != nil {
-		t.Fatalf("Error waiting for job: %v", err)
-	}
-
-	var stdout, stderr bytes.Buffer
-	command := exec.Command("kubectl", "exec", "test-secret-nested",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default",
-		"--",
-		"cat", fmt.Sprintf("/var/gcp-test-secrets/my/nested/path/%s", f.testSecretID))
-	command.Stdout = &stdout
-	command.Stderr = &stderr
-	if err := command.Run(); err != nil {
-		fmt.Println("Stdout:", stdout.String())
-		fmt.Println("Stderr:", stderr.String())
-		t.Fatalf("Could not read secret from container: %v", err)
-	}
-	if !bytes.Equal(stdout.Bytes(), []byte(f.testSecretID)) {
-		t.Fatalf("Secret value is %v, want: %v", stdout.String(), f.testSecretID)
-	}
-}
-
-func TestMountInvalidPath(t *testing.T) {
-	podFile := filepath.Join(f.tempDir, "test-invalid.yaml")
-	if err := replaceTemplate("templates/test-invalid.yaml.tmpl", podFile); err != nil {
-		t.Fatalf("Error replacing pod template: %v", err)
-	}
-
-	if err := execCmd(exec.Command("kubectl", "apply", "--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default", "-f", podFile)); err != nil {
-		t.Fatalf("Error creating job: %v", err)
-	}
-
-	// We cannot use a 'wait for condition' since we are expecting a failure (that gets retried indefinitely).
-	// Instead wait for enough time to give the kubelet a chance to attempt the mount and have it fail.
-	time.Sleep(15 * time.Second)
-
-	var stdout, stderr bytes.Buffer
-	command := exec.Command("kubectl", "get", "events", "--field-selector", "involvedObject.name=test-secret-invalid",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default")
-	command.Stdout = &stdout
-	command.Stderr = &stderr
-	if err := command.Run(); err != nil {
-		fmt.Println("Stdout:", stdout.String())
-		fmt.Println("Stderr:", stderr.String())
-		t.Fatalf("Could not read secret from container: %v", err)
-	}
-	if !strings.Contains(stdout.String(), "invalid path") {
-		t.Fatalf("Unable to find 'invalid path' error: %v", stdout.String())
-	}
-}
-
-func TestMountSyncSecret(t *testing.T) {
-	podFile := filepath.Join(f.tempDir, "test-sync.yaml")
-	if err := replaceTemplate("templates/test-sync.yaml.tmpl", podFile); err != nil {
-		t.Fatalf("Error replacing pod template: %v", err)
-	}
-
-	if err := execCmd(exec.Command(
-		"kubectl", "apply", "-f", podFile,
-		"--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default",
-	)); err != nil {
-		t.Fatalf("Error creating job: %v", err)
-	}
-
-	// As a workaround for https://github.com/kubernetes/kubernetes/issues/83242, we sleep to
-	// ensure that the job resources exists before attempting to wait for it.
-	time.Sleep(5 * time.Second)
-	if err := execCmd(exec.Command(
-		"kubectl", "wait", "pod/test-secret-mounter-sync",
-		"--for=condition=Ready",
-		"--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default",
-		"--timeout", "5m",
-	)); err != nil {
-		t.Fatalf("Error waiting for job: %v", err)
-	}
-
-	var stdout, stderr bytes.Buffer
-	command := exec.Command(
-		"kubectl", "exec", "test-secret-mounter-sync",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default",
-		"--",
-		"printenv",
-	)
-	command.Stdout = &stdout
-	command.Stderr = &stderr
-	if err := command.Run(); err != nil {
-		fmt.Println("Stdout:", stdout.String())
-		fmt.Println("Stderr:", stderr.String())
-		t.Fatalf("Could not read secret from container: %v", err)
-	}
-	if got := stdout.Bytes(); !bytes.Contains(got, []byte(f.testSecretID)) {
-		t.Fatalf("pod env value is %s, does not contain: %s", string(got), f.testSecretID)
-	}
-
-	// TODO: Add checks for regional secret
-}
-
-func TestMountRotateSecret(t *testing.T) {
-	secretA := []byte("secreta")
-	secretB := []byte("secretb")
-
-	// Enable rotation.
-	check(execCmd(exec.Command("enable-rotation.sh", f.kubeconfigFile)))
-
-	// Wait for deployment to finish.
-	time.Sleep(45 * time.Second)
-
-	// Create test secret.
-	secretFileA := filepath.Join(f.tempDir, "secretValue-A")
-	check(os.WriteFile(secretFileA, secretA, 0644))
-	check(execCmd(exec.Command(
-		"gcloud", "secrets", "create", f.testRotateSecretID,
-		"--replication-policy", "automatic",
-		"--data-file", secretFileA,
-		"--project", f.testProjectID,
-	)))
-
-	// create a regional test secret
-	check(execCmd(exec.Command("gcloud", "config", "set", "api_endpoint_overrides/secretmanager",
-		"https://secretmanager."+f.location+".rep.googleapis.com/")))
-
-	check(execCmd(exec.Command(
-		"gcloud", "secrets", "create", f.testRotateSecretID,
-		"--location", f.location,
-		"--data-file", secretFileA,
-		"--project", f.testProjectID,
-	)))
-	check(execCmd(exec.Command("gcloud", "config", "unset", "api_endpoint_overrides/secretmanager")))
-
-	// Deploy the test pod.
-	podFile := filepath.Join(f.tempDir, "test-rotate.yaml")
-	if err := replaceTemplate("templates/test-rotate.yaml.tmpl", podFile); err != nil {
-		t.Fatalf("Error replacing pod template: %v", err)
-	}
-
-	if err := execCmd(exec.Command("kubectl", "apply", "--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default", "-f", podFile)); err != nil {
-		t.Fatalf("Error creating job: %v", err)
-	}
-
-	// As a workaround for https://github.com/kubernetes/kubernetes/issues/83242, we sleep to
-	// ensure that the job resources exists before attempting to wait for it.
-	time.Sleep(5 * time.Second)
-	if err := execCmd(exec.Command(
-		"kubectl", "wait", "pod/test-secret-mounter-rotate",
-		"--for=condition=Ready",
-		"--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default",
-		"--timeout", "5m",
-	)); err != nil {
-		execCmd(exec.Command(
-			"kubectl", "describe", "pods",
-			"--namespace", "default",
-			"--kubeconfig", f.kubeconfigFile,
-		))
-		t.Fatalf("Error waiting for job: %v", err)
-	}
-
-	var stdout, stderr bytes.Buffer
-	command := exec.Command(
-		"kubectl", "exec", "test-secret-mounter-rotate",
-		"--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default",
-		"--",
-		"cat", "/var/gcp-test-secrets/rotate")
-	command.Stdout = &stdout
-	command.Stderr = &stderr
-	if err := command.Run(); err != nil {
-		fmt.Println("Stdout:", stdout.String())
-		fmt.Println("Stderr:", stderr.String())
-		t.Fatalf("Could not read secret from container: %v", err)
-	}
-	if got := stdout.Bytes(); !bytes.Equal(got, secretA) {
-		t.Fatalf("Secret value is %v, want: %v", got, secretA)
-	}
-
-	// Rotate the secret.
-	secretFileB := filepath.Join(f.tempDir, "secretValue-B")
-	check(os.WriteFile(secretFileB, secretB, 0644))
-	check(execCmd(exec.Command(
-		"gcloud", "secrets", "versions", "add", f.testRotateSecretID,
-		"--data-file", secretFileB,
-		"--project", f.testProjectID,
-	)))
-
-	// Rotate regional secret
-	check(execCmd(exec.Command("gcloud", "config", "set", "api_endpoint_overrides/secretmanager",
-		"https://secretmanager."+f.location+".rep.googleapis.com/")))
-	check(execCmd(exec.Command(
-		"gcloud", "secrets", "versions", "add", f.testRotateSecretID,
-		"--data-file", secretFileB,
-		"--project", f.testProjectID,
-		"--location", f.location,
-	)))
-
-	check(execCmd(exec.Command("gcloud", "config", "unset", "api_endpoint_overrides/secretmanager")))
-
-	// Wait for the global secret to have 2 versions.
-	waitForMinVersions(t, f.testRotateSecretID, f.testProjectID, "" /* global */, 2, 180*time.Second)
-
-	time.Sleep(150 * time.Second)
-
-	// Verify update.
-	stdout.Reset()
-	stderr.Reset()
-	command = exec.Command(
-		"kubectl", "exec", "test-secret-mounter-rotate",
-		"--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default",
-		"--",
-		"cat", "/var/gcp-test-secrets/rotate",
-	)
-	command.Stdout = &stdout
-	command.Stderr = &stderr
-	if err := command.Run(); err != nil {
-		fmt.Println("Stdout:", stdout.String())
-		fmt.Println("Stderr:", stderr.String())
-		t.Fatalf("Could not read secret from container: %v", err)
-	}
-	if got := stdout.Bytes(); !bytes.Equal(got, secretB) {
-		t.Fatalf("Secret value is %v, want: %v", got, secretB)
-	}
-}
-
-// Execute a test job that mounts a extract secret and checks that the value is correct.
-func TestMountExtractSecret(t *testing.T) {
-	secretData := []byte(`{"user":"admin", "password":"password@1234"}`)
-
-	// Create test secret
-	secretFile := filepath.Join(f.tempDir, "secretExtractValue")
-	check(os.WriteFile(secretFile, secretData, 0644))
-	check(execCmd(exec.Command(
-		"gcloud", "secrets", "create", f.testExtractSecretID,
-		"--replication-policy", "automatic",
-		"--data-file", secretFile,
-		"--project", f.testProjectID,
-	)))
-
-	podFile := filepath.Join(f.tempDir, "test-extract-key.yaml")
-	if err := replaceTemplate("templates/test-extract-key.yaml.tmpl", podFile); err != nil {
-		t.Fatalf("Error replacing pod template: %v", err)
-	}
-
-	if err := execCmd(exec.Command("kubectl", "apply", "--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default", "-f", podFile)); err != nil {
-		t.Fatalf("Error creating job: %v", err)
-	}
-
-	if err := execCmd(exec.Command("kubectl", "wait", "pod/test-secret-mounter-extract", "--for=condition=Ready",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default", "--timeout", "5m")); err != nil {
-		t.Fatalf("Error waiting for job: %v", err)
-	}
-	testExtractSecret := []byte("admin")
-
-	// Check Mounted Secrets
-	var stdout, stderr bytes.Buffer
-	command := exec.Command(
-		"kubectl", "exec", "test-secret-mounter-extract",
-		"--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default",
-		"--",
-		"cat", "/var/gcp-test-secrets/extract")
-	command.Stdout = &stdout
-	command.Stderr = &stderr
-	if err := command.Run(); err != nil {
-		fmt.Println("Stdout:", stdout.String())
-		fmt.Println("Stderr:", stderr.String())
-		t.Fatalf("Could not read secret from container: %v", err)
-	}
-	if got := stdout.Bytes(); !bytes.Equal(got, testExtractSecret) {
-		t.Fatalf("Secret value is %v, want: %v", got, testExtractSecret)
-	}
-}
-
-// mounts global and regional parameter versions and checks whether they are equivalent or not (both json and yaml)
-func TestMountParameterVersion(t *testing.T) {
-	podFile := filepath.Join(f.tempDir, "test-parameter-version-pod.yaml")
-	if err := replaceTemplate("templates/test-parameter-version-pod.yaml.tmpl", podFile); err != nil {
-		t.Fatalf("Error replacing pod template: %v", err)
-	}
-
-	if err := execCmd(exec.Command("kubectl", "apply", "--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default", "-f", podFile)); err != nil {
-		t.Fatalf("Error creating job: %v", err)
-	}
-
-	// As a workaround for https://github.com/kubernetes/kubernetes/issues/83242, we sleep to
-	// ensure that the job resources exists before attempting to wait for it.
-	time.Sleep(5 * time.Second)
-	if err := execCmd(exec.Command("kubectl", "wait", "pod/test-parameter-version-mounter", "--for=condition=Ready",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default", "--timeout", "5m")); err != nil {
-		t.Fatalf("Error waiting for job: %v", err)
-	}
-
-	if err := checkMountedParameterVersion(
-		"test-parameter-version-mounter", // podName
-		fmt.Sprintf("/var/gcp-test-parameter-version/%s/global/%s", f.parameterIdYaml, f.parameterVersionIdYAML), // mounted file path
-		fmt.Sprintf("user: admin\nuser2: support\ndb_pwd: %s\n", f.testSecretID),                                 // expected payload
-	); err != nil {
-		t.Fatalf("Error while testing global yaml parameter version: %v", err)
-	}
-
-	if err := checkMountedParameterVersion(
-		"test-parameter-version-mounter", // podName
-		fmt.Sprintf("/var/gcp-test-parameter-version/%s/global/%s", f.parameterIdJson, f.parameterVersionIdJSON), // mounted filepath
-		fmt.Sprintf("{\"user\": \"admin\",\n \"user2\": \"support\", \"db_pwd\": \"%s\"}", f.testSecretID),       // expected payload
-	); err != nil {
-		t.Fatalf("Error while testing global json parameter version: %v", err)
-	}
-
-	if err := checkMountedParameterVersion(
-		"test-parameter-version-mounter", // podName
-		fmt.Sprintf("/var/gcp-test-parameter-version/%s/%s/%s", f.regionalParameterIdYAML, f.location, f.regionalParameterVersionIdYAML), // mounted filepath
-		fmt.Sprintf("user: admin\nuser2: support\ndb_pwd: %s-regional\n", f.testSecretID),                                                // expected payload
-	); err != nil {
-		t.Fatalf("Error while testing regional yaml parameter version: %v", err)
-	}
-
-	if err := checkMountedParameterVersion(
-		"test-parameter-version-mounter", // podName
-		fmt.Sprintf("/var/gcp-test-parameter-version/%s/%s/%s", f.regionalParameterIdJSON, f.location, f.regionalParameterVersionIdJSON), // filepath
-		fmt.Sprintf("{\"user\": \"admin\",\n \"user2\": \"support\", \"db_pwd\": \"%s-regional\"}", f.testSecretID),                      // expected payload
-	); err != nil {
-		t.Fatalf("Error while testing regional json parameter version: %v", err)
-	}
-}
-
-// mounts global and regional parameter versions and applies extractJSONKey whenever applicable
-func TestMountParameterVersionExtractKeys(t *testing.T) {
-	podFile := filepath.Join(f.tempDir, "test-parameter-version-extract-keys.yaml")
-	if err := replaceTemplate("templates/test-parameter-version-extract-keys.yaml.tmpl", podFile); err != nil {
-		t.Fatalf("Error replacing pod template: %v", err)
-	}
-
-	if err := execCmd(exec.Command("kubectl", "apply", "--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default", "-f", podFile)); err != nil {
-		t.Fatalf("Error creating job: %v", err)
-	}
-
-	// As a workaround for https://github.com/kubernetes/kubernetes/issues/83242, we sleep to
-	// ensure that the job resources exists before attempting to wait for it.
-	time.Sleep(5 * time.Second)
-	if err := execCmd(exec.Command("kubectl", "wait", "pod/test-parameter-version-key-extraction", "--for=condition=Ready",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default", "--timeout", "5m")); err != nil {
-		t.Fatalf("Error waiting for pod test-parameter-version-key-extraction: %v", err)
-	}
-
-	if err := checkMountedParameterVersion(
-		"test-parameter-version-key-extraction", // podName
-		fmt.Sprintf("/var/gcp-test-parameter-version-keys/%s/global/%s", f.parameterIdYaml, f.parameterVersionIdYAML), // mounted file path
-		fmt.Sprintf("user: admin\nuser2: support\ndb_pwd: %s\n", f.testSecretID),                                      // expected payload
-	); err != nil {
-		t.Fatalf("Error while testing global yaml parameter version extracted key 'db_pwd': %v", err) // expected global secret
-	}
-
-	if err := checkMountedParameterVersion(
-		"test-parameter-version-key-extraction", // podName
-		fmt.Sprintf("/var/gcp-test-parameter-version-keys/%s/global/%s", f.parameterIdJson, f.parameterVersionIdJSON), // mounted filepath
-		"admin", // expected payload (extractJSONKey with key user used)
-	); err != nil {
-		t.Fatalf("Error while testing global json parameter version extracted key 'user': %v", err)
-	}
-
-	if err := checkMountedParameterVersion(
-		"test-parameter-version-key-extraction", // podName
-		fmt.Sprintf("/var/gcp-test-parameter-version-keys/%s/%s/%s", f.regionalParameterIdYAML, f.location, f.regionalParameterVersionIdYAML), // mounted filepath
-		fmt.Sprintf("user: admin\nuser2: support\ndb_pwd: %s-regional\n", f.testSecretID),                                                     // expected payload
-	); err != nil {
-		t.Fatalf("Error while testing regional yaml parameter version extracted key 'user2': %v", err)
-	}
-
-	if err := checkMountedParameterVersion(
-		"test-parameter-version-key-extraction", // podName
-		fmt.Sprintf("/var/gcp-test-parameter-version-keys/%s/%s/%s", f.regionalParameterIdJSON, f.location, f.regionalParameterVersionIdJSON), // filepath
-		fmt.Sprintf("%s-regional", f.testSecretID), // expected payload (extractJSONKey used with key db_pwd used)
-	); err != nil {
-		t.Fatalf("Error while testing regional json parameter version extracted key 'db_pwd': %v", err) // expected regional secret
-	}
-}
-
-// mounts global and regional yaml and json parameter versions at the exact ..data locations, not at their symlinks
-func TestMountParameterVersionFileMode(t *testing.T) {
-	podFile := filepath.Join(f.tempDir, "test-parameter-version-pod-mode.yaml")
-	if err := replaceTemplate("templates/test-parameter-version-pod-mode.yaml.tmpl", podFile); err != nil {
-		t.Fatalf("Error replacing pod template: %v", err)
-	}
-
-	if err := execCmd(exec.Command("kubectl", "apply", "--kubeconfig", f.kubeconfigFile,
-		"--namespace", "default", "-f", podFile)); err != nil {
-		t.Fatalf("Error creating job: %v", err)
-	}
-
-	// As a workaround for https://github.com/kubernetes/kubernetes/issues/83242, we sleep to
-	// ensure that the job resources exists before attempting to wait for it.
-	time.Sleep(5 * time.Second)
-	if err := execCmd(exec.Command("kubectl", "wait", "pod/test-parameter-version-mounter-filemode", "--for=condition=Ready",
-		"--kubeconfig", f.kubeconfigFile, "--namespace", "default", "--timeout", "5m")); err != nil {
-		t.Fatalf("Error waiting for pod test-parameter-version-mounter-filemode: %v", err)
-	}
-
-	if err := checkMountedParameterVersionFileMode(
-		fmt.Sprintf("/var/gcp-test-parameter-version-mode/..data/%s/global/%s", f.parameterIdYaml, f.parameterVersionIdYAML), // mounted file path
-		"420", // expected mode
-	); err != nil {
-		t.Fatalf("Error while testing global yaml parameter version: %v", err)
-	}
-
-	if err := checkMountedParameterVersionFileMode(
-		fmt.Sprintf("/var/gcp-test-parameter-version-mode/..data/%s/global/%s", f.parameterIdJson, f.parameterVersionIdJSON), // mounted filepath
-		"600", // expected mode
-	); err != nil {
-		t.Fatalf("Error while testing global json parameter version: %v", err)
-	}
-
-	if err := checkMountedParameterVersionFileMode(
-		fmt.Sprintf("/var/gcp-test-parameter-version-mode/..data/%s/%s/%s", f.regionalParameterIdYAML, f.location, f.regionalParameterVersionIdYAML), // mounted filepath
-		"400", // expected mode
-	); err != nil {
-		t.Fatalf("Error while testing regional yaml parameter version filemode: %v", err)
-	}
-
-	if err := checkMountedParameterVersionFileMode(
-		fmt.Sprintf("/var/gcp-test-parameter-version-mode/..data/%s/%s/%s", f.regionalParameterIdJSON, f.location, f.regionalParameterVersionIdJSON), // filepath
-		"440", // expected mode
-	); err != nil {
-		t.Fatalf("Error while testing regional json parameter version filemode: %v", err)
-	}
 }
